node/agents/gemini_strategist/pyproject.toml
[project]
name = "heiwa-gemini-strategist"
version = "0.1.0"
description = "The central intelligence agent for Heiwa mesh."
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.7.0",
    "google-generativeai>=0.4.1",
    "pydantic>=2.0.0"
]
